# docs/api.yaml
openapi: '3.0.0'
info:
  title: 'Recommendations HOTS API'
  description: 'Work with user management, published events, recommendations'
  termsOfService: 'TODO'
  version: '1.0'
  contact:
    url: 'TODO'
    name: 'Recommendations HOTS API'
    email: 'api@example.com'
  license:
    name: 'MIT'
servers:
  - url: http://localhost:3000/v1
    description: Development Server
paths:
...
  /auth/activate/{code}:
    get:
      security: []
      operationId: activateUser
      tags:
        - auth
      summary: Activates a newly created user
      description: >
        After creation of a new user, the user has to be activated by supplying the
        activation code sent out by the system. After activation the user can be
        logged in to the system.
      parameters:
        - in: path
          name: code
          schema:
            type: string
          required: true
      responses:
        204:
          $ref: '#/components/responses/NoContent'
        400:
          $ref: '#/components/responses/InvalidRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    InvalidRequestError:
      type: object
      properties:
        errors:
          type: object
  responses:
    InvalidRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InvalidRequestError'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: An unkown error has occured, please retry later.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NoContent:
      description: No content
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth:
      - admin
      - user